BUG: Use uint16 as output in otbImageClassifier
The otbImageClassifier application seems to be written with the
intention of outputting uint16 images, but there is a typo in the code
that sets the output image pixel type. Because of this, when the output
contains a class with an identifier larger than 255, it gets saturated
and written as 255.

I have not run the tests due to some problems with them on this
machine. It might be necessary to update the expected outputs.

Signed-off-by: Laurentiu Nicola <lnicola@dend.ro>