PERF: Clear our sample lists before RF training
The random forests machine learning model gets a list of training
samples which it has to adapt to a cv::Mat as required by OpenCV. OTB's
training samples are not used afterwards, so we can clear the lists to
free some RAM.

On a larger scene this change reduced the memory amount during training
from 22.7 GB to 20.2 GB, which seems about right. Most of the memory is
actually used by the OpenCV model (and again becomes problematic during
classification), but there's not much to do there. Still, for us this
was a welcome improvement.